Desktop/api_automation/requirements.txt
# HTTP and API Testing
requests
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-cov==4.1.0
pytest-json-report==1.5.0

# Data Handling
jsonschema==4.21.1
PyYAML==6.0.1
python-dotenv==1.0.1

# Reporting and Logging
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-json-report==1.5.0

# Utilities
Faker==22.6.0
deepdiff==6.7.1

python-dotenv
